requirements.txt
APScheduler==3.11.2
solders==0.22.0
base58==2.1.1
websockets==13.1
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.2
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
